Third Person Controller- WORKING
The third person controller that's included in the Standard Assets
actually works if you attach a camera tagged "Main Camera" to it. This
works best with the ThirdPersonCamera prefabs that come with the asset.
It will auto target the character if its tagged "Player," but its best
to just drop the controller into the target feild. The animator for an
external model will work if you select that model's avatar on the
ThirdPersonController's inspector. I think I will push this branch soon,
even if the other third person conroller asset changed some project
settings. I don't think it changed too much.